+++ examples/data-sources/aiven_influxdb/data-source.tf
@@ -1,4 +1,0 @@
-data "aiven_influxdb" "inf1" {
-  project      = data.aiven_project.pr1.project
-  service_name = "my-inf1"
-}