+++ examples/resources/aiven_influxdb/import.sh
@@ -1,1 +1,0 @@
-terraform import aiven_influxdb.inf1 project/service_name