+++ internal/sdkprovider/service/influxdb/sweep.go
@@ -1,9 +1,0 @@
-package influxdb
-
-import (
-	"github.com/aiven/terraform-provider-aiven/internal/sweep"
-)
-
-func init() {
-	sweep.AddServiceSweeper("influxdb")
-}